requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.10.1
aiortc==1.9.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncua==1.1.6
attrs==25.3.0
av==12.3.0
backoff==2.2.1
boto3==1.35.60
botocore==1.35.99
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
exceptiongroup==1.3.0
fastapi==0.110.3
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.1
google-crc32c==1.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
humanfriendly==9.2
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
inference==0.50.1
iniconfig==2.1.0
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
nvidia-ml-py==12.575.51
onnxruntime-gpu==1.22.0
openai==1.82.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
pillow==11.2.1
pluggy==1.6.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.22.0
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
py-cpuinfo==9.0.0
pybase64==1.0.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.7.1
pydantic_core==2.33.2
pydot==2.0.0
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pylogix==1.0.5
pymodbus==3.8.3
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyreadline==2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
s3transfer==0.10.4
scikit-image==0.25.2
scipy==1.15.3
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
slack_sdk==3.33.5
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
structlog==24.4.0
supervision==0.25.1
sympy==1.14.0
tifffile==2025.5.10
tldextract==5.1.3
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
torchvision==0.22.0+cu128
tqdm==4.67.1
twilio==9.3.8
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.146
ultralytics-thop==2.0.14
urllib3==2.4.0
wait_for2==0.3.2
yarl==1.20.0
zxing-cpp==2.2.0